refactor: move functionsManifest to separate file
